+++ Plutarch/Prelude.hs
@@ -51,6 +51,8 @@ module Plutarch.Prelude (
   pnot,
   (#&&),
   (#||),
+  pcond,
+  potherwise,
 
   -- * Bytestrings and bytestring utilities
 
@@ -223,9 +225,11 @@ import Plutarch.Builtin (
  )
 import Plutarch.Builtin.Bool (
   PBool (PFalse, PTrue),
+  pcond,
   pif,
   pif',
   pnot,
+  potherwise,
   (#&&),
   (#||),
  )