+++ CHANGELOG.md
@@ -35,6 +35,8 @@
 * `Plutarch.Internal.Eq` module
 * `Plutarch.Internal.Ord` module
 * `pif'` to `Plutarch.Prelude`
+* `pcond` as a Plutarch equivalent to multi-way if
+* `potherwise`, with a similar purpose to `otherwise` in Haskell
 
 ## Changed